car_order/requirements.txt
Django==4.1.2
python-dotenv==0.21.0
psycopg2-binary==2.9.3
djangorestframework==3.14.0